+++ mobile/lib/modules/asset_viewer/ui/video_player.dart
@@ -43,3 +43,4 @@ class VideoPlayerViewer extends HookConsumerWidget {
     );
   }
 }
+
