+++ mobile/lib/modules/search/ui/curated_people_row.dart
@@ -7,6 +7,7 @@ import 'package:immich_mobile/utils/image_url_builder.dart';
 
 class CuratedPeopleRow extends StatelessWidget {
   final List<CuratedContent> content;
+  final EdgeInsets? padding;
 
   /// Callback with the content and the index when tapped
   final Function(CuratedContent, int)? onTap;
@@ -16,6 +17,7 @@ class CuratedPeopleRow extends StatelessWidget {
     super.key,
     required this.content,
     this.onTap,
+    this.padding,
     required this.onNameTap,
   });
 
@@ -43,6 +45,7 @@ class CuratedPeopleRow extends StatelessWidget {
     }
 
     return ListView.builder(
+      padding: padding,
       scrollDirection: Axis.horizontal,
       itemBuilder: (context, index) {
         final person = content[index];
